# api/doc/swagger/comment.json
{
  "swagger": "2.0",
  "info": {
    "title": "Comment API",
    "version": "1.0.0"
  },
  "paths": {
    "/posts/{id}/comments": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "format": "int64",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of comments",
            "schema": {
              "items": {
                "$ref": "#/definitions/Comment"
              },
              "type": "array"
            }
          }
        },
        "tags": [
          "comment"
        ],
        "operationId": "listComments",
        "summary": "List all comments for a blog post"
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "format": "in64",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewComment"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Comment added"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Post not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "comment"
        ],
        "operationId": "addComment",
        "summary": "Add a comment to a blog post"
      }
    }
  },
  "definitions": {
    "Comment": {
      "properties": {
        "author_id": {
          "format": "int64",
          "type": "integer"
        },
        "content": {
          "type": "string"
        },
        "created_at": {
          "format": "date-time",
          "type": "string"
        },
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "post_id": {
          "format": "int64",
          "type": "integer"
        }
      },
      "type": "object"
    },
    "NewComment": {
      "properties": {
        "content": {
          "type": "string"
        }
      },
      "required": [
        "content"
      ],
      "type": "object"
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "tags": [
    {
      "description": "Operations related to comments on posts",
      "name": "comment"
    }
  ],
  "x-components": {}
}