# internal/app/comment/port/genhttp/openapi_types.gen.go
// Package genhttp provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package genhttp

import (
	"time"
)

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// Comment defines model for Comment.
type Comment struct {
	AuthorId  *int64     `json:"author_id,omitempty"`
	Content   *string    `json:"content,omitempty"`
	CreatedAt *time.Time `json:"created_at,omitempty"`
	Id        *int64     `json:"id,omitempty"`
	PostId    *int64     `json:"post_id,omitempty"`
}

// NewComment defines model for NewComment.
type NewComment struct {
	Content string `json:"content"`
}

// AddCommentJSONRequestBody defines body for AddComment for application/json ContentType.
type AddCommentJSONRequestBody = NewComment